Optimized implementation of Eq and Ord.
Based on the benchmarks, comparing equal strings of 4 or 10 bytes was
pretty much equivalent to comparing strings of 20 bytes, since an empty
slice was still built.

Now, we avoid doing this if the string is stored fully in the prefix.
In the case where it is inlined, we compare the last 8 bytes of the
struct as-is, without building a slice.